Refactored suction grip and added more tests
