+++ configs/tasks/rearrange/pick.yaml
@@ -104,6 +104,9 @@ TASK:
             DISABLE_GRIP: False
             DELTA_POS_LIMIT: 0.0125
             EE_CTRL_LIM: 0.015
+            # Suction grasp params
+            SUCTION_THRESH_DIST: 0.05
+            SUCTION_OFFSET: [0.0, 0.0, 0.0]
     POSSIBLE_ACTIONS:
         - ARM_ACTION
